Rename packages to mvalceleanu for unique package artifactory
