+++ app/build.gradle
@@ -6,7 +6,7 @@ android {
     compileSdkVersion project.compileSdk
 
     defaultConfig {
-        applicationId "pub.devrel.easypermissions.sample"
+        applicationId "mvalceleanu.easypermissions.sample"
         minSdkVersion 14
         targetSdkVersion targetSdk
         versionCode 1
@@ -25,5 +25,5 @@ android {
 dependencies {
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
     implementation 'androidx.appcompat:appcompat:1.2.0'
-    implementation project(':easypermissions')
+    implementation project(':easypermissions-ktx')
 }
