+++ easypermissions-ktx/bintray.gradle
@@ -1,0 +1,121 @@
+apply plugin: 'com.jfrog.bintray'
+apply plugin: 'maven-publish'
+
+group = mavenGroup
+version = mavenVersion
+
+task sourcesJar(type: Jar) {
+    from android.sourceSets.main.java.srcDirs
+    archiveClassifier.set("sources")
+}
+
+task javadoc(type: Javadoc) {
+    source = android.sourceSets.main.java.srcDirs
+    classpath += configurations.compile
+    classpath += project.files(android.getBootClasspath().join(File.pathSeparator))
+
+    failOnError = false
+}
+
+task javadocJar(type: Jar, dependsOn: javadoc) {
+    archiveClassifier.set("sources")
+    from javadoc.destinationDir
+}
+
+artifacts {
+    archives sourcesJar
+    archives javadocJar
+}
+
+publishing {
+    publications {
+        library(MavenPublication) {
+
+            groupId mavenGroup
+            artifactId mavenArtifactId
+            version mavenVersion
+
+            artifact "$buildDir/outputs/aar/$mavenArtifactId-release.aar"
+            artifact javadocJar
+            artifact sourcesJar
+
+            pom.withXml {
+                // Name
+                asNode().appendNode('name', "EasyPermissions-ktx")
+
+                // Description
+                asNode().appendNode('description', 'A Kotlin wrapper library for basic Android M system permissions logic')
+
+                // URL
+                asNode().appendNode('url', 'https://github.com/VMadalin/easypermissions-ktx')
+
+                // Inception Year
+                asNode().appendNode('inceptionYear', '2020')
+
+                // Licenses
+                def license = asNode().appendNode('licenses').appendNode('license')
+                license.appendNode('name', 'The Apache Software License, Version 2.0')
+                license.appendNode('url', 'http://www.apache.org/licenses/LICENSE-2.0.txt')
+                license.appendNode('distribution', 'repo')
+
+                // Developers
+                def developer = asNode().appendNode('developers').appendNode('developer')
+                developer.appendNode('name', 'Madalin Valceleanu')
+
+                // SCM
+                def scm = asNode().appendNode('scm')
+                scm.appendNode('connection', 'https://github.com/VMadalin/easypermissions-ktx.git')
+                scm.appendNode('url', 'https://github.com/VMadalin/easypermissions-ktx')
+
+                // Dependencies
+                def dependenciesNode = asNode()['dependencies'][0]
+                if (dependenciesNode == null) {
+                    dependenciesNode = asNode().appendNode('dependencies')
+                }
+
+                // Add all that are 'compile'
+                configurations.api.allDependencies.each {
+                    def dependencyNode = dependenciesNode.appendNode('dependency')
+                    dependencyNode.appendNode('groupId', it.group)
+                    dependencyNode.appendNode('artifactId', it.name)
+                    dependencyNode.appendNode('version', it.version)
+                }
+            }
+        }
+    }
+}
+
+bintray {
+    user = hasProperty('BINTRAY_USER') ? getProperty('BINTRAY_USER') : System.getenv('BINTRAY_USER')
+    key = hasProperty('BINTRAY_KEY') ? getProperty('BINTRAY_KEY') : System.getenv('BINTRAY_KEY')
+
+    filesSpec {
+        from "$buildDir/publications/library/pom-default.xml"
+        into "mvalceleanu/$mavenArtifactId/$mavenVersion/"
+        rename { String fileName ->
+            "${mavenArtifactId}-${mavenVersion}.pom"
+        }
+    }
+
+    configurations = [ 'archives' ]
+
+    pkg {
+        repo = projectName
+        name = mavenArtifactId
+        userOrg = bintrayOrg
+        licenses = [ 'Apache-2.0' ]
+        vcsUrl = "${githubUrl}.git"
+
+        version {
+            name = mavenVersion
+            released = new Date()
+        }
+    }
+}
+
+afterEvaluate { project ->
+    def pomTask = "generatePomFileForLibraryPublication"
+
+    // Convenience task to prepare everything we need for releases
+    task prepareArtifacts(dependsOn: [javadocJar, sourcesJar, assembleRelease, pomTask]) {}
+}
