+++ easypermissions-ktx/build.gradle
@@ -1,0 +1,75 @@
+apply plugin: 'com.android.library'
+apply plugin: 'kotlin-android-extensions'
+apply plugin: 'kotlin-android'
+apply plugin: 'kotlin-allopen'
+apply plugin: 'com.vanniktech.android.junit.jacoco'
+
+allOpen {
+    annotation("mvalceleanu.easypermissions.annotations.Mockable")
+}
+
+android {
+    compileSdkVersion project.compileSdk
+
+    defaultConfig {
+        minSdkVersion 14
+        targetSdkVersion targetSdk
+        versionCode 1
+        versionName mavenVersion
+        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+    }
+
+    buildTypes {
+        debug {
+            testCoverageEnabled true
+        }
+        release {
+            minifyEnabled false
+            consumerProguardFiles 'proguard-rules.pro'
+        }
+    }
+
+    compileOptions {
+        sourceCompatibility = JavaVersion.VERSION_1_8
+        targetCompatibility = JavaVersion.VERSION_1_8
+    }
+
+    kotlinOptions {
+        jvmTarget = "1.8"
+    }
+
+    testOptions {
+        unitTests {
+            includeAndroidResources = true
+        }
+    }
+}
+
+junitJacoco {
+    jacocoVersion = '0.8.4'
+    includeNoLocationClasses = true
+}
+
+dependencies {
+    api "androidx.appcompat:appcompat:1.2.0"
+    api "androidx.core:core-ktx:1.3.2"
+    api "androidx.fragment:fragment-ktx:1.2.5"
+
+    implementation "androidx.core:core-ktx:1.3.2"
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+
+    testImplementation 'junit:junit:4.13.1'
+    testImplementation 'com.google.truth:truth:1.0'
+    testImplementation 'androidx.test:core:1.3.0'
+    testImplementation 'androidx.fragment:fragment-testing:1.2.5'
+    testImplementation 'org.robolectric:robolectric:4.4'
+    testImplementation 'org.mockito:mockito-core:3.0.0'
+    testImplementation 'com.nhaarman.mockitokotlin2:mockito-kotlin:2.1.0'
+
+    testAnnotationProcessor 'com.google.auto.service:auto-service:1.0-rc4'
+}
+
+apply from: 'bintray.gradle'
+repositories {
+    mavenCentral()
+}
