+++ easypermissions-ktx/src/main/AndroidManifest.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--suppress AndroidDomInspection -->
+<manifest
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    package="mvalceleanu.easypermissions">
+
+    <application>
+        <activity android:name=".components.TestActivity"/>
+        <activity android:name=".components.TestAppCompatActivity"/>
+        <activity android:name=".components.TestSupportFragmentActivity"/>
+    </application>
+
+</manifest>
