+++ easypermissions-ktx/src/main/res/values/colors.xml
