+++ easypermissions-ktx/src/main/res/values/strings.xml
