+++ easypermissions-ktx/src/main/res/values/styles.xml
