+++ easypermissions-ktx/src/test/java/pub/devrel/easypermissions/components/TestActivity.kt
@@ -1,0 +1,43 @@
+/*
+ * Copyright 2019 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package mvalceleanu.easypermissions.components
+
+import android.app.Activity
+import mvalceleanu.easypermissions.annotations.AfterPermissionGranted
+import mvalceleanu.easypermissions.annotations.Mockable
+import mvalceleanu.easypermissions.facade.EasyPermissions
+
+@Mockable
+class TestActivity :
+    Activity(),
+    EasyPermissions.PermissionCallbacks,
+    EasyPermissions.RationaleCallbacks {
+
+    companion object {
+        const val REQUEST_CODE = 1
+    }
+
+    override fun onPermissionsGranted(requestCode: Int, perms: List<String>) {}
+
+    override fun onPermissionsDenied(requestCode: Int, perms: List<String>) {}
+
+    override fun onRationaleAccepted(requestCode: Int) {}
+
+    override fun onRationaleDenied(requestCode: Int) {}
+
+    @AfterPermissionGranted(REQUEST_CODE)
+    fun afterPermissionGranted() {}
+}
